missed the re-type to int on the Q_INVOKABLE method for the qml

+++ Shared/alerts/AlertConditionsController.cpp
@@ -1154,7 +1154,7 @@ bool AlertConditionsController::conditionAlreadyAdded(const QString& conditionNa
   return false;
 }
 
-void AlertConditionsController::setSelectedTargetNameIndex(qint16 currentIndex)
+void AlertConditionsController::setSelectedTargetNameIndex(int currentIndex)
 {
   m_selectedTargetNameIndex = currentIndex;
 }
